backend/requirements.txt

fastapi
uvicorn
pydantic
pymongo
pytest
pytest-asyncio
httpx
python-dotenv
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]